requirements.txt
aiogram==3.26.0
python-dotenv==1.2.2
gTTS==2.5.4
deep-translator==1.11.4